LayoutTests/webrtc/audio-replace-track.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>Testing local audio capture playback causes "playing" event to fire</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src ="routines.js"></script>
    <script>
    promise_test((test) => {
        if (window.testRunner)
            testRunner.setUserMediaPermission(true);

        var sender;
        var firsStream;
        var secondStream;
        var remoteStream;
        return navigator.mediaDevices.getUserMedia({ audio: { sampleRate: { exact: 48000 } } }).then((stream) => {
            firstStream = stream;
            return navigator.mediaDevices.getUserMedia({ audio: { sampleRate: { exact: 48000 } } });
        }).then((stream) => {
            secondStream = stream;
            if (window.internals)
                internals.useMockRTCPeerConnectionFactory("TwoRealPeerConnections");
            return new Promise((resolve, reject) => {
                createConnections((firstConnection) => {
                    sender = firstConnection.addTrack(firstStream.getAudioTracks()[0], firstStream);
                }, (secondConnection) => {
                    secondConnection.ontrack = (trackEvent) => { resolve(trackEvent.streams[0]); };
                });
                setTimeout(() => reject("Test timed out"), 5000);
            });
        }).then((stream) => {
            remoteStream = stream;
            return analyseAudio(remoteStream, 1000);
        }).then((results) => {
            assert_true(results.heardHum, "heard hum");
            return sender.replaceTrack(secondStream.getAudioTracks()[0], secondStream);
        }).then(() => {
            assert_true(sender.track === secondStream.getAudioTracks()[0]);
            return waitFor(500);
        }).then((results) => {
            return analyseAudio(remoteStream, 1000);
        }).then((results) => {
            assert_true(results.heardHum, "heard hum");
        });
    }, "Replacing audio track from a peer connection");
    </script>
</head>
<body>
</body>
</html>
